Rename bot.py to client.py
